Reorganize Channel IDs to reflect guac changes.
Now channels stored in memory does not have Ids. They are hidden behind
a "state" which can contain Address (if channel is New - only registered as
a counteparty) and later when the channel is created on chain it will
change its state to Open which now will contain its ChannelId.

Changing ID will unify states a channel can be in. Before it was not
clear enough what state a channel is in: channel_id.is_none(),
channel_id.is_some() and channel_state which could contain any of 5 enum
values. With this change channels in memory are still uniquely
identified by state (we can have channel with only a single address,
and once its on the chain its still unique with ChannelId), and also
their intention is clean.

There is a little bit to optimize here: now they are stored in a Vec,
which makes the lookups linear, nonetheless it could become a HashMap
with a tricky update of the Key whenever a state changes in the channel
structure.